Databázové systémy/Zadania 2, 3, 5 a 6/myproject/requirements.txt
Django~=3.2.12
psycopg2~=2.9.3
djangorestframework~=3.13.1
requests~=2.27.1
SQLAlchemy~=1.4.31